+++ features/fixtures/app/main.go
@@ -45,8 +45,6 @@ func configureBasicBugsnag(testcase string) {
 	}
 
 	switch testcase {
-	case "endpoint-legacy":
-		config.Endpoint = os.Getenv("BUGSNAG_ENDPOINT")
 	case "endpoint-notify":
 		config.Endpoints = bugsnag.Endpoints{Notify: os.Getenv("BUGSNAG_ENDPOINT")}
 	case "endpoint-session":